Refactor AbstractSps class and fix a method
The AbstractSps class has been refactored to improve the handling of various database platforms. The platform name identifying method has been updated to provide broader support. Changes also include the optimisation of code handling the "executeQuery" and "execute" methods. Furthermore, a bug in the setSqlFunction method in the Rule class has been corrected to bind the closure to the correct class context.